Handle HTTP Refresh header
