+++ components/script/dom/htmlmetaelement.rs
@@ -2,15 +2,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use std::str::FromStr;
-use std::sync::LazyLock;
-use std::time::Duration;
-
 use constellation_traits::NavigationHistoryBehavior;
 use dom_struct::dom_struct;
 use html5ever::{LocalName, Prefix, local_name, namespace_url, ns};
 use js::rust::HandleObject;
-use regex::bytes::Regex;
 use servo_url::ServoUrl;
 use style::str::HTML_SPACE_CHARACTERS;
 
@@ -30,7 +25,6 @@ use crate::dom::node::{BindContext, Node, NodeTraits, UnbindContext};
 use crate::dom::virtualmethods::VirtualMethods;
 use crate::dom::window::Window;
 use crate::script_runtime::CanGc;
-use crate::timers::OneshotTimerCallback;
 
 #[dom_struct]
 pub(crate) struct HTMLMetaElement {
@@ -160,87 +154,8 @@ impl HTMLMetaElement {
         // 1
         if !content.is_empty() {
             // 3
-            self.shared_declarative_refresh_steps(content);
-        }
-    }
-
-    /// <https://html.spec.whatwg.org/multipage/#shared-declarative-refresh-steps>
-    fn shared_declarative_refresh_steps(&self, content: DOMString) {
-        // 1
-        let document = self.owner_document();
-        if document.will_declaratively_refresh() {
-            return;
-        }
-
-        // 2-11
-        static REFRESH_REGEX: LazyLock<Regex> = LazyLock::new(|| {
-            Regex::new(
-                r#"(?x)
-                ^
-                \s* # 3
-                ((?<time>\d+)\.?|\.) # 5-6
-                [0-9.]* # 8
-                (
-                    (;|,| ) # 10.1
-                    \s* # 10.2
-                    (;|,)? # 10.3
-                    \s* # 10.4
-                    (
-                        (U|u)(R|r)(L|l) # 11.2-11.4
-                        \s*=\s* # 11.5-11.7
-                        ('(?<url1>.*?)'?|"(?<url2>.*?)"?|(?<url3>[^'"].*)) # 11.8 - 11.10
-                        |
-                        (?<url4>.*)
-                    )?
-                )?
-                $
-            "#,
-            )
-            .unwrap()
-        });
-
-        let mut url_record = document.url();
-        let captures = if let Some(captures) = REFRESH_REGEX.captures(content.as_bytes()) {
-            captures
-        } else {
-            return;
-        };
-        let time = if let Some(time_string) = captures.name("time") {
-            u64::from_str(&String::from_utf8_lossy(time_string.as_bytes())).unwrap_or(0)
-        } else {
-            0
-        };
-        let captured_url = captures.name("url1").or(captures
-            .name("url2")
-            .or(captures.name("url3").or(captures.name("url4"))));
-
-        if let Some(url_match) = captured_url {
-            url_record = if let Ok(url) = ServoUrl::parse_with_base(
-                Some(&url_record),
-                &String::from_utf8_lossy(url_match.as_bytes()),
-            ) {
-                url
-            } else {
-                return;
-            }
-        }
-        // 12-13
-        if document.completely_loaded() {
-            // TODO: handle active sandboxing flag
-            let window = self.owner_window();
-            window.as_global_scope().schedule_callback(
-                OneshotTimerCallback::RefreshRedirectDue(RefreshRedirectDue {
-                    window: window.clone(),
-                    url: url_record,
-                }),
-                Duration::from_secs(time),
-            );
-            document.set_declarative_refresh(DeclarativeRefresh::CreatedAfterLoad);
-        } else {
-            document.set_declarative_refresh(DeclarativeRefresh::PendingLoad {
-                url: url_record,
-                time,
-            });
+            self.owner_document()
+                .shared_declarative_refresh_steps(content);
         }
     }
 }
